pyproject.toml
[tool.poetry]
name = "user-auth-system"
version = "0.1.0"
description = "User Authentication and Management System using Django"
authors = ["OpenHands <openhands@all-hands.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
